+++ api/v1alpha1/userconfig/service/kafka/zz_generated.deepcopy.go
@@ -614,6 +614,16 @@ func (in *KafkaUserConfig) DeepCopyInto(out *KafkaUserConfig) {
 		*out = new(bool)
 		**out = **in
 	}
+	if in.BackupIntervalHours != nil {
+		in, out := &in.BackupIntervalHours, &out.BackupIntervalHours
+		*out = new(int)
+		**out = **in
+	}
+	if in.BackupRetentionDays != nil {
+		in, out := &in.BackupRetentionDays, &out.BackupRetentionDays
+		*out = new(int)
+		**out = **in
+	}
 	if in.CustomDomain != nil {
 		in, out := &in.CustomDomain, &out.CustomDomain
 		*out = new(string)